requirements.txt
asgiref==3.2.10
Django==3.0.8
djangorestframework==3.11.0
psycopg2-binary==2.8.5
pytz==2020.1
sqlparse==0.3.1
django_heroku==0.3.1
django_filter==2.3.0
gunicorn==20.0.4
dj_database_url==0.5.0
django-nested-admin==3.3.2
